Clean Up and improve greenhouse climate system

+++ src/main/java/forestry/api/climate/IClimatiserDefinition.java
@@ -9,7 +9,7 @@ public interface IClimatiserDefinition {
 
 	float getChange();
 
-	double getRange();
+	float getRange();
 
 	EnumClimatiserModes getMode();
