+++ src/main/java/forestry/arboriculture/items/ItemBlockWoodDoor.java
@@ -38,7 +38,7 @@ public class ItemBlockWoodDoor extends ItemBlockWood<BlockArbDoor> {
 			ItemStack itemstack = player.getHeldItem(hand);
 
 			if (player.canPlayerEdit(pos, facing, itemstack) && this.block.canPlaceBlockAt(worldIn, pos)) {
-				EnumFacing enumfacing = EnumFacing.fromAngle((double) player.rotationYaw);
+				EnumFacing enumfacing = EnumFacing.fromAngle(player.rotationYaw);
 				int i = enumfacing.getFrontOffsetX();
 				int j = enumfacing.getFrontOffsetZ();
 				boolean flag = i < 0 && hitZ < 0.5F || i > 0 && hitZ > 0.5F || j < 0 && hitX > 0.5F || j > 0 && hitX < 0.5F;
